+++ CHANGELOG.md
@@ -33,6 +33,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Deprecated
 - Deprecate `api.LatencyImpactMarkdownTable` in favor of `api.impact.LatencyImpactMarkdownTable`.
 
+### Fixed
+- Sort impact tables by action alphabetically.
+
 ## [3.15.0] - 2023-07-03
 [3.15.0]: https://github.com/atlassian/report/compare/release-3.14.0...release-3.15.0